# internal/infra/http/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gym API
  version: 1.0.0
...
paths:
...
  /exercises:
    get:
      tags: [Exercises]
      summary: List exercises
      description: |
        Returns a cursor-paginated list of exercises. Use `?search=` to filter by
        name (case-insensitive). Pass the `cursor.next_cursor` value from the
        previous response as `?cursor=` to fetch the next page.
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Filter by name (partial, case-insensitive)
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque cursor from a previous response to fetch the next page
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: Paginated list of exercises
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExercisePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Exercises]
      summary: Create exercise
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExerciseRequest'
      responses:
        '201':
          description: Exercise created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /exercises/{id}:
    get:
      tags: [Exercises]
      summary: Get exercise
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Tasks ───────────────────────────────────────────────────────────────────
...
  /workouts:
    get:
      tags: [Workouts]
      summary: List user workouts
      description: Returns all workouts belonging to the authenticated user, each including its exercises.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of workouts with exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkoutDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Workouts]
      summary: Create workout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkoutRequest'
      responses:
        '201':
          description: Workout created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /workouts/{id}:
    get:
      tags: [Workouts]
      summary: Get workout with exercises
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Workout with exercises
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Workouts]
      summary: Update workout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkoutRequest'
      responses:
        '200':
          description: Updated workout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Workouts]
      summary: Delete workout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Workout deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workouts/{id}/exercises:
    post:
      tags: [Workouts]
      summary: Add exercise to workout
      description: Adds an exercise to the workout. A workout can have at most 50 exercises.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWorkoutExerciseRequest'
      responses:
        '201':
          description: Exercise added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutExercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workouts/{id}/exercises/reorder:
    patch:
      tags: [Workouts]
      summary: Reorder workout exercises
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderWorkoutExercisesRequest'
      responses:
        '204':
          description: Exercises reordered
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workouts/{id}/exercises/{weId}:
    put:
      tags: [Workouts]
      summary: Update workout exercise
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: weId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkoutExerciseRequest'
      responses:
        '200':
          description: Updated workout exercise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutExercise'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Workouts]
      summary: Remove exercise from workout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: weId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Exercise removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workouts/{id}/progress:
    get:
      tags: [Workouts]
      summary: Get exercise evolution for a workout
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: exercise_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter progress to a single exercise
      responses:
        '200':
          description: Progress per exercise
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExerciseProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Workout Sessions ──────────────────────────────────────────────────────────
  /workout-sessions/missed:
    get:
      tags: [WorkoutSessions]
      summary: List missed workout sessions
      description: |
        Returns days where a workout was scheduled but no session (complete or
        incomplete) was recorded. Bonus sessions (unscheduled days) are excluded.
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: Start date (YYYY-MM-DD). Defaults to 30 days ago.
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: End date (YYYY-MM-DD). Defaults to today.
      responses:
        '200':
          description: List of missed sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MissedSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /workout-sessions/{id}:
    get:
      tags: [WorkoutSessions]
      summary: Get session with recorded sets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSessionDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [WorkoutSessions]
      summary: Update session status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkoutSessionRequest'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkoutSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ExercisePage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Exercise'
        cursor:
          $ref: '#/components/schemas/CursorMeta'

    # ── Exercise schemas ──────────────────────────────────────────────────────
    Exercise:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Bench Press
        type:
          type: string
          enum: [repetition, time]
        unit:
          type: string
          example: kg
        created_at:
          type: string
          format: date-time
    CreateExerciseRequest:
      type: object
      required: [name, type, unit]
      properties:
        name:
          type: string
          example: Bench Press
        type:
          type: string
          enum: [repetition, time]
        unit:
          type: string
          example: kg

    # ── Task schemas ──────────────────────────────────────────────────────────
...
    WorkoutDetail:
      allOf:
        - $ref: '#/components/schemas/Workout'
        - type: object
          properties:
            exercises:
              type: array
              items:
                $ref: '#/components/schemas/WorkoutExercise'
            done_today:
              type: boolean
              description: Whether the workout already has a complete session recorded for the current day.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...